Add desktop and metainfo file for Flatpak
While I was in the process of submitting the Flatpak under the
new app ID, a reviewer noted that these files should be stored
in upstream repository. So we do that here.

The metainfo file should be updated as we release a new version;
the first `<release>` tag in `<releases>` should be updated
as we make changes to the app.

Signed-off-by: rany <ranygh@riseup.net>